feat: Complete migration planning and add layer stubs
- Complete migration planning docs (Phases 1-7): layer map, session
  design-specs, view/input design-specs, cross-cutting plan, meta-game
  plan, deletion sequence, implementation checklist
- Add Session layer stubs: IGameSession, CommandRouter, LocalSession,
  HostSession, ClientSession, NetworkSession, IdentityMap, SnapshotCodec
- Add View layer stubs: BattleScene, EntityManager, UnitVisual,
  ProjectileVisual, SummonerVisual (via prior commit)
- Add Input layer stub: InputCollector
- Extract BehaviorState and Team enums to standalone files
- Archive outdated rewrite-research docs, reorganize doc structure
- Remove unused Abilities/ system (BaseAbility, SlowOnHitAbility)
- Minor fixes: SimBehavior test updates, fire_spider scene fix

+++ scripts/csharp/Multiplayer/Sync/StateSnapshotBuilder.cs
@@ -150,7 +150,7 @@ public class StateSnapshotBuilder
                 targetNetworkId,
                 unit.IsAlive,
                 unit.ActivationState,
-                unit.BehaviorState,
+                (int)unit.BehaviorState,
                 unit.IsFacingRight
             ));
         }
